Permit listening to a subest of projects
For Heroku Review apps, we're trying to use a small database plan (with limited
number of collections allowed) and a small number of workers processing tasks
in order to reduce the cost of the app.

Using the PROJECTS_TO_INGEST variable we can define a small number of repos to
pay attention to and discard all other tasks.

You can remove PROJECTS_TO_INGEST to go back to using a normal ingestion behaviour.